int->NSInteger to remove warnings

+++ fe-aqua/ChannelListWin.mm
@@ -473,14 +473,14 @@ strip_crap (const char *s)
 
 // Table View Data Source Methods
 
-- (int) numberOfRowsInTableView:(NSTableView *) aTableView
+- (NSInteger) numberOfRowsInTableView:(NSTableView *) aTableView
 {
     return [items count];
 }
 
 - (id) tableView:(NSTableView *) aTableView
     objectValueForTableColumn:(NSTableColumn *) aTableColumn
-    row:(int) rowIndex
+    row:(NSInteger) rowIndex
 {
     one_entry *e = [items objectAtIndex:rowIndex];
     
@@ -494,7 +494,7 @@ strip_crap (const char *s)
     return @"";
 }
 
-- (BOOL)tableView:(NSTableView *)aTableView shouldEditTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (BOOL)tableView:(NSTableView *)aTableView shouldEditTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
     return NO;
 }
