+++ fe-aqua/LogViewer.mm
@@ -265,14 +265,14 @@ extern "C" {
 //
 // Delegate Methods
 
-- (int) numberOfRowsInTableView:(NSTableView *) aTableView
+- (NSInteger) numberOfRowsInTableView:(NSTableView *) aTableView
 {
     return [my_items count];
 }
 
 - (id) tableView:(NSTableView *) aTableView
     objectValueForTableColumn:(NSTableColumn *) aTableColumn
-    row:(int) rowIndex
+    row:(NSInteger) rowIndex
 {
     OneLog *item = [my_items objectAtIndex:rowIndex];
