+++ fe-aqua/NotifyListWin.mm
@@ -204,14 +204,14 @@ extern "C" {
 //////////////
 //
 
-- (int) numberOfRowsInTableView:(NSTableView *) aTableView
+- (NSInteger) numberOfRowsInTableView:(NSTableView *) aTableView
 {
     return [my_items count];
 }
 
 - (id) tableView:(NSTableView *) aTableView
     objectValueForTableColumn:(NSTableColumn *) aTableColumn
-    row:(int) rowIndex
+    row:(NSInteger) rowIndex
 {
     oneNotify *item = [my_items objectAtIndex:rowIndex];
